backend/requirements.txt
# Precision Engine and Dyno, LLC - Backend Requirements
# Unified Platform for E-commerce + Auto Machining Services

# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation & Serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Task Queue
celery==5.3.6
redis==5.0.1

# Email
sendgrid==6.11.0
aiosmtplib==3.0.1
jinja2==3.1.3

# SMS
twilio==8.13.0

# File Storage
boto3==1.34.37
python-magic==0.4.27

# PDF Generation
weasyprint==61.0
reportlab==4.0.9

# Date/Time
python-dateutil==2.8.2
pytz==2024.1

# CSV/Excel
openpyxl==3.1.2
pandas==2.2.0

# Image Processing
Pillow==10.2.0

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3
structlog==24.1.0

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0

# Database Migrations
alembic==1.13.1
